dom/imported-tests/webapps/WebStorage/tests/submissions/Ms2ger/test_storage_local_clear_js.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Web Storage</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
    <h1>storage_local_clear</h1>
    <div id="log"></div>
    <script>
        test(function() {
            localStorage.clear();
            localStorage.setItem("name", "user1");

            assert_not_equals(localStorage.name, undefined);
            assert_equals(localStorage.length, 1);

            localStorage.clear();
            assert_equals(localStorage.name, undefined, "localStorage.name")
            assert_equals(localStorage.length, 0, "localStorage.length")
        });
    </script>
 </body>
</html>
